Fix editor states

+++ src/components/posts/editor/resources/Resources.tsx
@@ -1,6 +1,6 @@
 import { EditorResource } from "@/pages/posts/[type]/[postId]/editor";
-import Resource from "@/components/posts/editor/resources/Resource";
-import { useCallback, useState } from "react";
+import Resource, { ResourceStates } from "@/components/posts/editor/resources/Resource";
+import { useCallback, useState, useTransition } from "react";
 import { useDropzone, DropEvent, FileRejection } from "react-dropzone";
 import FakeResource from "./FakeResource";
 import ensureLength from "@/utils/ensureLength";
@@ -16,15 +16,24 @@ interface ResourcesProps {
 }
 
 const Resources = ({ resources, setResources, postId, setMarkdown }: ResourcesProps) => {
+    const [mutatingResources, mutateResourcesTransition] = useTransition();
+    const mutateResources = useCallback((resources: EditorResource[]) => {
+        mutateResourcesTransition(() => setResources(resources));
+    }, []);
+
     const addResource = useCallback(
         async (acceptedFiles: File[], fileRejections: FileRejection[], event: DropEvent) => {
-            let newResources = [];
+            if (mutatingResources) return;
+
+            const newResources: EditorResource[] = [];
             for (const file of acceptedFiles) {
                 let [filename, filetype] = file.name.split(".");
+
+                // Doesn't include file extension
                 const resourceName = sanitize(filename).replace("%20", "_").replace(/\s/g, "_");
 
-                const makeId = (number: number) =>
-                    `${resourceName}_${ensureLength(String(number), 4)}`;
+                const makeId = (counter: number) =>
+                    `${resourceName}_${ensureLength(String(counter), 4)}`;
 
                 let resourceNumber = 1;
                 while (
@@ -41,34 +50,38 @@ const Resources = ({ resources, setResources, postId, setMarkdown }: ResourcesPr
                 filename = `${resourceId}.${filetype}`;
 
                 const newResource = {
-                    ref: resourceId,
+                    id: resourceId,
                     title: resourceName,
                     filename: filename,
                     type: file.type.split("/")[0],
                     description: "",
-                    url: resourceUrl(filename),
+
                     data: ((await toB64(file)) as string).replace(/^data:image\/\w+;base64,/, ""),
                     mimetype: file.type,
                     postId: postId,
                 };
-                fetch(`/api/resources/${resourceId}`, {
+
+                await fetch(`/api/resources/${resourceId}`, {
                     method: "POST",
                     headers: {
                         "Content-Type": "application/json",
                     },
                     body: JSON.stringify(newResource),
-                });
-
-                newResources.push(newResource);
+                })
+                    .then((resp) => {
+                        if (resp.ok)
+                            newResources.push({ ...newResource, url: resourceUrl(filename) });
+                    })
+                    .catch((e) => console.log(e));
             }
 
-            setResources(
+            mutateResources(
                 [...resources, ...newResources].filter(
                     (resource) => resource !== null && resource !== undefined
                 )
             );
         },
-        []
+        [resources]
     );
 
     const { getRootProps, getInputProps, isDragActive } = useDropzone({
@@ -76,46 +89,69 @@ const Resources = ({ resources, setResources, postId, setMarkdown }: ResourcesPr
     });
 
     const removeResource = useCallback(
-        async (index: number) => {
-            const resource = resources[index];
-            const resp = await fetch(`/api/resources/${resource.ref}`, {
-                method: "DELETE",
-            });
+        (index: number) => {
+            if (mutatingResources) return;
 
-            if (resp.ok || resp.status === 404) {
-                const slicedResources = [...resources];
-                delete slicedResources[index];
-                setResources(slicedResources);
-            }
+            fetch(`/api/resources/${resources[index].id}`, {
+                method: "DELETE",
+            }).then((resp) => {
+                if (resp.ok || resp.status === 404) {
+                    const slicedResources = [...resources];
+                    delete slicedResources[index];
+                    mutateResources(slicedResources);
+                }
+            });
         },
         [resources]
     );
 
-    const pushUpdate = useCallback(() => {
-        fetch(`/api/posts/${postId}/resources`)
-            .then((resp) => resp.json())
-            .then((resp) => {
-                return resp.resources || [];
-            })
-            .then(setResources);
-    }, []);
+    const updateResource = useCallback(
+        async (index: number, newResource: EditorResource) => {
+            const oldResource = resources[index];
+            console.log(oldResource, newResource, resources, index);
+            if (mutatingResources) return oldResource;
+
+            await fetch(`/api/resources/${oldResource.id}`, {
+                method: "PUT",
+                headers: {
+                    "Content-Type": "application/json",
+                },
+                body: JSON.stringify(newResource),
+            }).then((resp) => {
+                if (resp.ok) {
+                    const newResources = [...resources];
+                    newResources[index] = newResource;
+
+                    mutateResources(newResources);
+                }
+            });
+            return oldResource;
+        },
+        [resources]
+    );
 
     return (
-        <div className={`h-full ${isDragActive ? "brightness-90" : "brightness-100"}`}>
+        <div className="h-full">
             <div className="grid grid-cols-2 gap-3 mt-4 justify-stretch relative">
                 {resources.map((resource, index) => {
                     return (
                         <Resource
+                            index={index}
                             remove={() => removeResource(index)}
                             resource={resource}
-                            pushUpdate={pushUpdate}
+                            updateResource={updateResource}
                             setMarkdown={setMarkdown}
                             postId={postId}
                             key={index}
                         />
                     );
                 })}
-                <div {...getRootProps()} className="relative cursor-pointer">
+                <div
+                    {...getRootProps()}
+                    className={`relative cursor-pointer ${
+                        isDragActive ? "brightness-90" : "brightness-100"
+                    }`}
+                >
                     <FakeResource placeholderId={null}></FakeResource>
                 </div>
             </div>
